Break out Event types and functions into own file
